MdePkg/Library: RISC-V: Add RAS agent client library
Add the RAS agent client library which is used to communicate to
a remote RAS agent. The RAS agent can be queried about the details
of various hardware error sources in the system.

There are systems which contain RAS drivers within Management Mode
environments. For these systems, the way to communicate RAS messages
is different in that it sends messages over MM protocol instead of
direct MPXY comms. In order to maintain common driver for both modes
we are introducing a function pointer based method to carry out this
communication based on a platform configurable PCD.
PcdMMPassThroughEnable==True enables driver to bypass MM mode.

Signed-off-by: Himanshu Chauhan <hchauhan@ventanamicro.com>
Signed-off-by: Dhaval Sharma <dhaval@rivosinc.com>